Explicitly use namespace:: before function calls in the code.
Prior to this change a set of calls to library() were made in a code
chunk which was not visible to the website viewers.  As a result the
code that was visible would not evaluate without the end user guessing
at the namespaces required.

This change removes the chunk with the library calls and explictly uses
the <namespace>::<function> calls.